Refactor discarded helpers into separate file
